# swagger/swagger.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: JR CMS
  description: This is the api server for JR CMS
  contact:
    name: Mason
    email: mason.xiong@yahoo.com
servers:
  - url: /v1
...
paths:
  /users:
    post:
      tags:
        - 'users'
      summary: 'Register a new user'
      description: ''
      operationId: 'addUser'
      requestBody:
        description: 'Required user fields'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        400:
          description: 'Bad Request'
        201:
          description: 'Created'
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: 'object'
      properties:
        email:
          type: 'string'
        password:
          type: 'string'
        name:
          type: 'string'
          description: 'display name'
...